Fix compiler warnings and needlessly failing catch tests. (#3171)

+++ arangod/Cache/PlainBucket.cpp
@@ -64,7 +64,7 @@ bool PlainBucket::isFull() const {
   return !hasEmptySlot;
 }
 
-CachedValue* PlainBucket::find(uint32_t hash, void const* key, uint32_t keySize,
+CachedValue* PlainBucket::find(uint32_t hash, void const* key, size_t keySize,
                                bool moveToFront) {
   TRI_ASSERT(isLocked());
   CachedValue* result = nullptr;
@@ -102,7 +102,7 @@ void PlainBucket::insert(uint32_t hash, CachedValue* value) {
 }
 
 CachedValue* PlainBucket::remove(uint32_t hash, void const* key,
-                                 uint32_t keySize) {
+                                 size_t keySize) {
   TRI_ASSERT(isLocked());
   CachedValue* value = find(hash, key, keySize, false);
   if (value != nullptr) {
